feat: field style support font-size\font-weight\font-style
