+++ packages/core/client/src/schema-settings/LinkageRules/LinkageRuleAction.tsx
@@ -203,96 +203,3 @@ export const FormButtonLinkageRuleAction = observer(
   },
   { displayName: 'FormButtonLinkageRuleAction' },
 );
-
-const colorSchema = {
-  type: 'string',
-  'x-decorator': 'FormItem',
-  'x-component': 'ColorPicker',
-  'x-component-props': {
-    defaultValue: '',
-  },
-};
-const textAlignSchema = {
-  type: 'string',
-  'x-decorator': 'FormItem',
-  'x-component': 'Select',
-  'x-component-props': {
-    defaultValue: '',
-  },
-  enum: [
-    {
-      label: 'left',
-      value: 'left',
-    },
-    {
-      label: 'right',
-      value: 'right',
-    },
-    {
-      label: 'center',
-      value: 'center',
-    },
-  ],
-};
-
-const schemas = new Map();
-schemas.set(ActionType.Color, colorSchema);
-schemas.set(ActionType.BackgroundColor, colorSchema);
-schemas.set(ActionType.TextAlign, textAlignSchema);
-
-export const FormStyleLinkageRuleAction = observer(
-  (props: any) => {
-    const { options, collectionName } = props;
-    const { t } = useTranslation();
-    const compile = useCompile();
-    const remove = useContext(RemoveActionContext);
-    const { operator, setOperator, value: fieldValue, setValue } = useValues(options);
-    const operators = useMemo(
-      () =>
-        compile([
-          { label: t('Color'), value: ActionType.Color, schema: {} },
-          { label: t('Background Color'), value: ActionType.BackgroundColor, schema: {} },
-          { label: t('Text Align'), value: ActionType.TextAlign, schema: {} },
-        ]),
-      [compile, t],
-    );
-
-    const onChange = useCallback(
-      (value) => {
-        setOperator(value);
-      },
-      [setOperator],
-    );
-
-    const closeStyle = useMemo(() => ({ color: '#bfbfbf' }), []);
-    return (
-      <div style={{ marginBottom: 8 }}>
-        <Space>
-          <Select
-            data-testid="select-linkage-properties"
-            popupMatchSelectWidth={false}
-            value={operator}
-            options={operators}
-            onChange={onChange}
-            placeholder={t('action')}
-          />
-          {[ActionType.Color, ActionType.BackgroundColor, ActionType.TextAlign].includes(operator) && (
-            <ValueDynamicComponent
-              fieldValue={fieldValue}
-              schema={schemas.get(operator)}
-              setValue={setValue}
-              collectionName={collectionName}
-              inputModes={['constant']}
-            />
-          )}
-          {!props.disabled && (
-            <a role="button" aria-label="icon-close">
-              <CloseCircleOutlined onClick={remove} style={closeStyle} />
-            </a>
-          )}
-        </Space>
-      </div>
-    );
-  },
-  { displayName: 'FormStyleLinkageRuleAction' },
-);
